# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JffEhr.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/MedshackTreatments": {
      "get": {
        "tags": [
          "MedshackTreatments"
        ],
        "parameters": [
          {
            "name": "visitId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedshackTreatmentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedshackTreatmentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedshackTreatmentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MedshackTreatments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMedshackTreatmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMedshackTreatmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMedshackTreatmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackTreatmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackTreatmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackTreatmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/MedshackTreatments/{id}": {
      "get": {
        "tags": [
          "MedshackTreatments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackTreatmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackTreatmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackTreatmentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MedshackTreatments"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/MedshackVisits": {
      "get": {
        "tags": [
          "MedshackVisits"
        ],
        "parameters": [
          {
            "name": "registrationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedshackVisitDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedshackVisitDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MedshackVisitDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MedshackVisits"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMedshackVisitRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMedshackVisitRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMedshackVisitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackVisitDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackVisitDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackVisitDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/MedshackVisits/{id}": {
      "get": {
        "tags": [
          "MedshackVisits"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackVisitDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackVisitDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedshackVisitDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "MedshackVisits"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMedshackTreatmentRequest": {
        "required": [
          "treatmentDescription",
          "treatmentTime",
          "visitId"
        ],
        "type": "object",
        "properties": {
          "visitId": {
            "type": "string",
            "format": "uuid"
          },
          "treatmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "treatmentDescription": {
            "minLength": 1,
            "type": "string"
          },
          "outcome": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateMedshackVisitRequest": {
        "required": [
          "reason",
          "registrationId",
          "visitAt"
        ],
        "type": "object",
        "properties": {
          "registrationId": {
            "type": "string",
            "format": "uuid"
          },
          "visitAt": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "minLength": 1,
            "type": "string"
          },
          "accompaniedBy": {
            "maxLength": 160,
            "type": "string",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "pulse": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bloodPressure": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "oxygenSaturation": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "medicalHistory": {
            "type": "string",
            "nullable": true
          },
          "signsSymptoms": {
            "type": "string",
            "nullable": true
          },
          "findings": {
            "type": "string",
            "nullable": true
          },
          "nursingReport": {
            "type": "string",
            "nullable": true
          },
          "adviceGiven": {
            "type": "string",
            "nullable": true
          },
          "doctorId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MedshackTreatmentDto": {
        "type": "object",
        "properties": {
          "treatmentId": {
            "type": "string",
            "format": "uuid"
          },
          "visitId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNo": {
            "type": "integer",
            "format": "int32"
          },
          "treatmentTime": {
            "type": "string",
            "format": "date-time"
          },
          "treatmentDescription": {
            "type": "string",
            "nullable": true
          },
          "outcome": {
            "type": "string",
            "nullable": true
          },
          "administeredBy": {
            "type": "string",
            "format": "uuid"
          },
          "administeredByName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MedshackVisitDto": {
        "type": "object",
        "properties": {
          "visitId": {
            "type": "string",
            "format": "uuid"
          },
          "registrationId": {
            "type": "string",
            "format": "uuid"
          },
          "visitAt": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "accompaniedBy": {
            "type": "string",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "pulse": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bloodPressure": {
            "type": "string",
            "nullable": true
          },
          "oxygenSaturation": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "medicalHistory": {
            "type": "string",
            "nullable": true
          },
          "signsSymptoms": {
            "type": "string",
            "nullable": true
          },
          "findings": {
            "type": "string",
            "nullable": true
          },
          "nursingReport": {
            "type": "string",
            "nullable": true
          },
          "adviceGiven": {
            "type": "string",
            "nullable": true
          },
          "nurseId": {
            "type": "string",
            "format": "uuid"
          },
          "nurseName": {
            "type": "string",
            "nullable": true
          },
          "doctorId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "doctorName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}